Add MIDI Output functionality
This intercepts on commands of each synth and if MIDI output is
selected, instead sends the corresponding MIDI message. Right now has
hardcoded checks to find out which synth is specifically being
triggered to determine MIDI message parameters. In the future, if more
synths are added, will need to add a general mechanism to each synth
that specifies how it maps to MIDI programs.